rename files and remove sample classes
