+++ Api/Src/Repositories/User.ts
@@ -1,6 +1,6 @@
 import { Repository } from './Repository';
 
-export default class User extends Repository {
+export default class UserRepository extends Repository {
     async getByEmail(email: string) {
         return this.model.findOne({ email });
     }
