--- Api/Src/Services/SampleService.ts
@@ -1,5 +1,0 @@
-export default class SampleService {
-    getHello = async () => {
-        return 'Hello';
-    };
-}
