--- Api/Test/SampleTest.test.ts
@@ -1,6 +1,0 @@
-
-describe('Sample test ', () => {
-    test('sample test', () =>{
-        expect(2 + 2).toBe(4);
-    })
-})
